pyproject.toml
[project]
name = "rag-server"
dynamic = ["version"]
description = "SGR RAG Agent"
readme = "README.md"
authors = [
    {name = "Gerbylev"},
]
keywords = ["ai", "RAG", "reasoning", "schema-guided", "openai"]

requires-python = ">=3.13,<4.0"
dependencies = [
    "aiogram>=3.17.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.115.11",
    "fastmcp>=2.12.4",
    "graypy>=2.1.0",
    "httpx>=0.28.1",
    "jambo>=0.1.3.post2",
    "jinja2>=3.1.0",
    "openai>=2.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.10.1",
    "python-logging-loki>=0.3.1",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.15.1",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.39",
    "tavily-python>=0.3.0",
    "trafilatura>=1.6.0",
    "uvicorn>=0.34.0",
    "yoyo-migrations>=9.0.0",
]

[project.urls]
TGBOT = "https://t.me/rag_agent_by_suai_bot"

[dependency-groups]
dev = [
    "aiosqlite>=0.19.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.397",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.1",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 160
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["B008"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"


[tool.pytest.ini_options]
pythonpath = ["./src", "./tests"]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
log_cli = true
log_cli_level = "DEBUG"
addopts = "--cov=src --cov-report=term --cov-report=html:tests/htmlcov --html=tests/pytest-report/index.html"


[tool.pyright]
venvPath = "."
venv = ".venv"

